+++ app/src/main/java/com/snowflowerthon/snowman/ui/ArchiveFragment.kt
@@ -1,10 +1,10 @@
-package com.snowflowerthon.snowman.vote.category
+package com.snowflowerthon.snowman.ui
 
 import android.os.Bundle
+import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.fragment.app.Fragment
 import com.snowflowerthon.snowman.R
 
 // TODO: Rename parameter arguments, choose names that match
@@ -14,14 +14,14 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [ItemFragment.newInstance] factory method to
+ * Use the [ArchiveFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class ItemFragment : Fragment() {
+class ArchiveFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
-
+    // Initial commit comment
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         arguments?.let {
@@ -35,7 +35,7 @@ class ItemFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_item, container, false)
+        return inflater.inflate(R.layout.fragment_archive, container, false)
     }
 
     companion object {
@@ -45,12 +45,12 @@ class ItemFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment ItemFragment.
+         * @return A new instance of fragment ArchiveFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            ItemFragment().apply {
+            ArchiveFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
