+++ app/src/main/java/com/snowflowerthon/snowman/ui/LoginActivity.kt
@@ -1,7 +1,8 @@
-package com.snowflowerthon.snowman
+package com.snowflowerthon.snowman.ui
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.snowflowerthon.snowman.R
 
 class LoginActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
