+++ app/src/main/java/com/snowflowerthon/snowman/ui/ProfileFragment.kt
@@ -1,10 +1,11 @@
-package com.snowflowerthon.snowman
+package com.snowflowerthon.snowman.ui
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.snowflowerthon.snowman.R
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
