+++ app/src/main/java/com/snowflowerthon/snowman/ui/vote/ShardViewModel.kt
@@ -1,0 +1,8 @@
+package com.snowflowerthon.snowman.ui.vote
+
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+
+class SharedViewModel : ViewModel() {
+    val variableX = MutableLiveData<Boolean>()
+}
