+++ app/src/main/java/com/snowflowerthon/snowman/ui/vote/VoteFragment.kt
@@ -1,4 +1,4 @@
-package com.snowflowerthon.snowman.vote
+package com.snowflowerthon.snowman.ui.vote
 
 
 import android.os.Bundle
@@ -8,11 +8,14 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
 import androidx.fragment.app.Fragment
-import com.google.android.material.bottomsheet.BottomSheetDialogFragment
+import androidx.lifecycle.Observer
+import androidx.lifecycle.ViewModelProvider
 import com.google.android.material.tabs.TabLayoutMediator
+import com.snowflowerthon.snowman.R
+import com.snowflowerthon.snowman.data.RetrofitClient
+import com.snowflowerthon.snowman.data.apiService
 import com.snowflowerthon.snowman.databinding.FragmentVoteBinding
-import com.snowflowerthon.snowman.vote.category.ClothesFragment
-import com.snowflowerthon.snowman.vote.category.TabPagerAdapter
+import com.snowflowerthon.snowman.ui.vote.category.TabPagerAdapter
 
 interface PenguinInteractionListener {
     fun dressPenguinWithScarf()
@@ -23,6 +26,8 @@ class VoteFragment : Fragment(), PenguinInteractionListener {
 
     private var _binding: FragmentVoteBinding? = null
     private val binding get() = _binding!!
+
+    private lateinit var sharedViewModel: SharedViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -36,8 +41,6 @@ class VoteFragment : Fragment(), PenguinInteractionListener {
         super.onViewCreated(view, savedInstanceState)
 
 //        // 탭 레이아웃의 탭 이름을 정의합니다. 여기서는 "Tab 1"과 "Tab 2"로 설정하였습니다.
-        val aFragment = ClothesFragment()
-        aFragment.setInteractionListener(this) // VoteFragment의 인스턴스를 전달
 
 
         val tabNames = listOf("아우터", "아이템")
@@ -48,21 +51,25 @@ class VoteFragment : Fragment(), PenguinInteractionListener {
             tab.text = tabNames[position]
         }.attach()
 
-//        // AFragment를 뷰페이저에 추가
-//        val aFragment = ClothesFragment()
-//        aFragment.setInteractionListener(this) // VoteFragment의 인스턴스를 전달
-//
-//        val viewPager2 = binding.myPageviewPager
-//        val adapter = TabPagerAdapter(this, 2) // 예시로 탭이 2개라 가정
-//        adapter.addFragment(aFragment)
-//        // 다른 프래그먼트 추가
-//
-//        viewPager2.adapter = adapter
-//
-//        // 뷰페이저와 탭 레이아웃 연결
-//        TabLayoutMediator(binding.tabLayout, viewPager2) { tab, position ->
-//            tab.text = "Tab ${position + 1}"
-//        }.attach()
+
+        binding.btnCustomSave.setOnClickListener {
+            val retrofitAPI = RetrofitClient.getInstance().create(apiService::class.java)
+            retrofitAPI.voteClothes()
+        }
+
+        // SharedViewModel을 가져옵니다.
+        sharedViewModel = ViewModelProvider(requireActivity()).get(SharedViewModel::class.java)
+
+        // variableX를 감시하고, 값이 변경될 때마다 처리할 로직을 구현합니다.
+        sharedViewModel.variableX.observe(viewLifecycleOwner, Observer { newValue ->
+            // newValue로부터 필요한 작업을 수행합니다.
+            // 예: newValue가 true이면 어떤 동작을 수행하고, false이면 다른 동작을 수행
+
+            Log.d("VoteFragment","패딩을 눌러버림")
+            binding.ivCloth.setImageResource(R.drawable.img_coat)
+
+        })
+
     }
 
     override fun onDestroyView() {
