+++ app/src/main/res/layout/fragment_item.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".vote.category.ItemFragment">
+    tools:context=".ui.vote.category.ItemFragment">
 
 
     <GridLayout
