+++ app/src/main/res/layout/fragment_vote.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".ui.MainActivity">
 
     <ImageButton
         android:id="@+id/btn_custom_reset"
@@ -32,7 +32,7 @@
 
 
     <ImageView
-        android:id="@+id/imageView"
+        android:id="@+id/iv_base"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:src="@drawable/img_base"
@@ -41,6 +41,11 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
+    <ImageView
+        android:id="@+id/iv_cloth"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:src="@drawable/img_neat"/>
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="200dp"
